move delete session dialog into data session components
