+++ src/components/DataSession/DeleteSessionDialog.vue
